backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
google-generativeai==0.8.3
whitenoise==6.8.2
